Rename bitdotio.py -> _bitdotio.py
